async yargs handler, added force flag

+++ ts/schema/aggregator.ts
@@ -12,21 +12,34 @@ import {
 import chalk from "chalk";
 import fs from "node:fs";
 import path from "node:path";
+import readLineSync from "readline-sync";
 import { AggregatorSchema, JobSchema, pubKeyConverter, pubKeyReviver } from ".";
 import { toAccountString, toPermissionString, toUtf8 } from "../utils";
 import { findProjectRoot } from "../utils/findProjectRoot";
 
 export const saveAggregatorSchema = (
   aggregatorSchema: AggregatorSchema,
-  outFile: string
+  outFile: string,
+  force = false
 ): void => {
   const fullPath = path.join(findProjectRoot(), outFile);
-  fs.writeFileSync(
-    fullPath,
-    JSON.stringify(aggregatorSchema, pubKeyConverter, 2)
-  );
-  console.log(`Aggregator Schema: saved to ${chalk.green(fullPath)}`);
-  return;
+
+  if (fs.existsSync(fullPath)) {
+    if (force) {
+      console.log(`Aggregator Schema: already existed, skipping ${fullPath}`);
+    }
+    console.log(fullPath);
+    if (readLineSync.keyInYN("Do you want to overwrite this file?")) {
+      fs.writeFileSync(
+        fullPath,
+        JSON.stringify(aggregatorSchema, pubKeyConverter, 2)
+      );
+      console.log(`Aggregator Schema: saved to ${chalk.green(fullPath)}`);
+      return;
+    } else {
+      return;
+    }
+  }
 };
 
 export const loadAggregatorDefinition = (
